# Startup performance improvements