@Code-Ac 2022-01-22
I initially followed your TODO comments but as it seems, the removal of the loop fixed it. Got up to only about a second for it by downloading every single library there is. I see this as no longer an issue (you can remove the todos).