@Code-Ac 2022-01-22
Also a question that is not related to the pr:
Would you be interested in an update of the IntelliJ project files? Had to fiddle around for a while to get it all working. Otherwise I will just add them to my `.gitignore`.