@benfry 2022-01-22
Thanks for looking into it. 

Please first check whether it's `ge.registerFont(font)` or `GraphicsEnvironment.getLocalGraphicsEnvironment()` that's taking all the time. It may be the latter if it's the first time that AWT classes are being loaded by the JVM, which means the 2 seconds will just happen elsewhere. 

Also, most functions like this need to happen on the EDT, where it's safe to perform AWT operations. So I'm a little nervous that we can't just throw it in a non-EDT thread like this, but should make sure we're safe.

And one more, that font needs to be available for the rest of the UI. So we need to make sure it's loaded before anything else is in use, so if we make this change we're gonna have to ensure that none of the other UI elements that expect the font to be available are being initialized in the meantime.