@Code-Ac 2022-01-22
It is not the class loading that takes that long, but the registering process.

I don't really know about AWT and EDT in Java - What should be unsave when calling this code in a separate thread? Also this code is first executed from the Main Thread (constructor of Base) (is this isn't the EDT Thread right?) so I think it should be possible to export the function into its own thread.
I will research all uses of `GraphicsEnvironment.getAllFonts()` will tell you in a moment.

I thought you were only using the registered font when calling for all accessible fonts in the settings, else using this function to get it (because it returns the font).