@Code-Ac 2022-01-22
So the first use case is in the `CreateFont` class, which is unused beside `SampleComponent` which is completely unused.
The second usage is in the core module. Class `PFont` uses it in the `loadFonts` method whit two usages in public methods. One in `list`, which has no usages, and one called `findFont` used in the `PGraphics` class in it's `createFont` method. There the method is surrounded by a try-catch.
The third use is in the `Toolkit` class itself, in the `getMonoFontList` method. It is passed through another series of methods until it ends up in the `PreferenceFrame` class. We know where it is used there.